+++ tests/models/test_models_cpu.py
@@ -12,6 +12,6 @@ def test_fix_max_steps_cpu(max_epochs, accumulate_grad_batches, batch_size):
         max_epochs,
         accumulate_grad_batches,
         batch_size,
-        gpus=0,
+        devices=0,
         accelerator="cpu",
     )